added level select menu, added load selected level object
only thing that needs to be implemented is add level to build in correct order. level buttons correspond to index already, make level select scene be the last in build just for convenience on how its set up rn. code/implementation is ugly but i understand it so lmk if there are any problems. i have 6 levels in level select rn but adding more is easy.